IT WORKS! UU matter controller in 20mins.
The logic is flipped (uu generator requires high for OFF), but it works.
I fixed the logic, but realized that this still will starve most of the
things.

Will think about it tomorrow. The point is that it is trivial to make a
nice state now.  Furthermore, wait instruction works PERFECTLY as
expected.